Krishnamurthy Number
Krishnamurthy Number:- It is a number that is equal to the sum of the factorials of all its digits. For example, 1, 2, 145. Sometimes it is also called a Strong number or Peterson number.

Examples of Krishnamurthy Number

1! = 1
So, 1 is a Krishnamurthy number.

Similarly,
2! = 1*2 = 2
2 is also a Krishnamurthy number.

145
=> 1! + 4! + 5!
=> 1 + 24 + 120
=> 145
The sum of the factorial of individual digits is the same as the original number 145. Hence, 145 is a Krishnamurthy number.

Similarly,
40585
=> 4! + 0! + 5! + 8! + 5!
=> 24 +1 + 120 + 40320 + 120
=> 40585

Hence the numbers 1, 2, 145 and 40585 are Krishnamurthy number.

151
=> 1! + 5! + 1!
=> 1 + 120 + 1
=> 122
The sum of the factorial of each individual digits is 122, which is not the same as the original number 151. It is not a Krishnamurthy number.